phase-2/backend/requirements.txt
fastapi
sqlmodel
pyjwt
requests
python-dotenv
psycopg2-binary
uvicorn[standard]
